requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
anthropic>=0.25.0
openai>=1.66.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.0.0
pymupdf>=1.23.0
cloudscraper>=1.2.71
playwright>=1.40.0
gunicorn>=21.2.0
